Create a new uploader that clears all of the image caches when you re-upload an image.

+++ src/SymEdit/Bundle/MediaBundle/Upload/UploadManager.php
@@ -60,6 +60,8 @@ class UploadManager implements UploadManagerInterface
             return;
         }
 
-        $this->filesystem->delete($media->getPath());
+        if ($this->filesystem->has($media->getPath())) {
+            $this->filesystem->delete($media->getPath());
+        }
     }
 }
